🌈 style(move): Move generate_data.py to the root dir.
